% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatMCA-package.R
\docType{package}
\name{SpatMCA-package}
\alias{SpatMCA}
\alias{SpatMCA-package}
\title{SpatMCA package}
\description{
Core namespace declarations and shared imports for the SpatMCA package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://egpivo.github.io/SpatMCA/}
  \item \url{https://github.com/egpivo/SpatMCA}
  \item Report bugs at \url{https://github.com/egpivo/SpatMCA/issues}
}

}
\author{
\strong{Maintainer}: Wen-Ting Wang \email{egpivo@gmail.com} (\href{https://orcid.org/0000-0003-3051-7302}{ORCID})

Authors:
\itemize{
  \item Hsin-Cheng Huang \email{hchuang@stat.sinica.edu.tw} (\href{https://orcid.org/0000-0002-5613-349X}{ORCID})
}

}
\keyword{internal}
